Add setDesignDocPreviousVersionsForTest to ensure tests do not leak state globally by resetting back to the original value after testing
